admin new button

+++ src/main/java/bau/kokany/controller/FXMLController.java
@@ -1,24 +1,42 @@
 package bau.kokany.controller;
 
 import java.net.URL;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 import bau.kokany.model.Expert;
 import bau.kokany.model.ExpertDAO;
 import bau.kokany.model.JPAExpertDAO;
 import bau.kokany.model.StatusType;
+import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.geometry.Insets;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.GridPane;
+import javafx.scene.layout.VBox;
 
 public class FXMLController implements Initializable {
 
+    void refresh() throws Exception {
+        row.removeAll();
+        try(ExpertDAO eDAO = new JPAExpertDAO();) {
+            List<Expert> expertList = eDAO.getExperts();
+            for(Expert e : expertList) {
+                if(row.size() < expertList.size()) {
+                    row.add(e);
+                }
+            }
+        }
+
+    }
     //Főoldal
     @FXML
     private AnchorPane main_page;
@@ -68,14 +86,7 @@ public class FXMLController implements Initializable {
             main_page.setVisible(false);
             admin_page.setVisible(true);
 
-            try(ExpertDAO eDAO = new JPAExpertDAO();) {
-                List<Expert> expertList = eDAO.getExperts();
-                for(Expert e : expertList) {
-                    if(row.size() < expertList.size()) {
-                        row.add(e);
-                    }
-                }
-            }
+            refresh();
         }
     }
 
@@ -107,15 +118,24 @@ public class FXMLController implements Initializable {
     @FXML
     private TableColumn<Expert, String> lista_status1;
 
+    ObservableList<Expert> row = FXCollections.observableArrayList();
+
     @FXML
     void admin_delete_buttonPushed(ActionEvent event) throws Exception {
         Expert selectedItem = lista1.getSelectionModel().getSelectedItem();
 
-        try(ExpertDAO eDAO = new JPAExpertDAO();) {
-            eDAO.deleteExpert(selectedItem);
-            lista1.getItems().remove(selectedItem);
-        }
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Szakember Törlése");
+        alert.setHeaderText(null);
+        alert.setContentText("Biztosan törlöd?");
 
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.get() == ButtonType.OK){
+            try(ExpertDAO eDAO = new JPAExpertDAO();) {
+                eDAO.deleteExpert(selectedItem);
+                lista1.getItems().remove(selectedItem);
+            }
+        }
 
     }
 
@@ -126,6 +146,41 @@ public class FXMLController implements Initializable {
 
     @FXML
     void admin_new_buttonPushed(ActionEvent event) {
+        Expert expert = new Expert();
+        Dialog<Expert> dialog = new Dialog<>();
+        dialog.setTitle("Új Szakember Felvétele");
+        dialog.setHeaderText(null);
+        DialogPane dialogPane = dialog.getDialogPane();
+        dialogPane.getButtonTypes().addAll(ButtonType.OK);
+        TextField expert_name = new TextField("Név");
+        TextField expert_prof= new TextField("Foglalkozás");
+        ObservableList<StatusType> options =
+                FXCollections.observableArrayList(StatusType.values());
+        ComboBox<StatusType> comboBox = new ComboBox<>(options);
+        comboBox.getSelectionModel().selectFirst();
+        dialogPane.setContent(new VBox(8, expert_name, expert_prof, comboBox));
+        Platform.runLater(expert_name::requestFocus);
+        dialog.setResultConverter((ButtonType button) -> {
+            if (button == ButtonType.OK) {
+                expert.setName(expert_name.getText());
+                expert.setProfession(expert_prof.getText());
+                expert.setStatus(comboBox.getValue());
+            }
+            return expert;
+
+        });
+        Optional<Expert> optionalResult = dialog.showAndWait();
+        optionalResult.ifPresent((Expert results) -> {
+            try(ExpertDAO eDAO = new JPAExpertDAO();) {
+                eDAO.saveExpert(expert);
+                lista1.getItems().add(expert);
+            }
+            catch (Exception e) {
+                e.printStackTrace();
+            }
+
+        });
+
 
     }
 
@@ -190,21 +245,12 @@ public class FXMLController implements Initializable {
     @FXML
     private TableColumn<Expert, String> lista_status;
 
-    ObservableList<Expert> row = FXCollections.observableArrayList();
-
     @FXML
     private Button ListExpertButton;
 
     @FXML
     void ListExpertButtonPushed(ActionEvent event) throws Exception {
-        try(ExpertDAO eDAO = new JPAExpertDAO();) {
-            List<Expert> expertList = eDAO.getExperts();
-            for(Expert e : expertList) {
-                if(row.size() < expertList.size()) {
-                    row.add(e);
-                }
-            }
-        }
+        refresh();
     }
 
     @FXML
